Revert to avoid weird error
com.peregrine-cms.base.core presents the following:
org.apache.sling.api.resource,version=[2.10,2.11) -- Cannot be resolved

when we extend ResourceResolver. Why? No idea, but that's the wrong version resolved by bundle plugin...
